+++ ceph/run-setup-osd.sh
@@ -89,6 +89,11 @@ UUID=$(uuidgen)
 OSD_SECRET=$(ceph-authtool --gen-print-key)
 
 ID=$(echo "{\"cephx_secret\": \"$OSD_SECRET\"}" |  ceph osd new $UUID -i - -n osd. -k /tmp/ceph.keyring)
+if [ "x" = "x$ID" ]; then
+    echo "EXIT: NO OSD-ID !"
+    exit 1
+fi
+
 
 echo "re-create /var/lib/ceph/osd/ceph-$ID"
 rm -rf /var/lib/ceph/osd/ceph-$ID
@@ -109,5 +114,16 @@ systemctl start ceph-osd@$ID
 sleep 2
 
 systemctl status ceph-osd@$ID
+
+#touch /var/spool/cron/crontabs/root
+#chown root:crontab /var/spool/cron/crontabs/root
+#chmod 600 /var/spool/cron/crontabs/root
+echo "@reboot sleep 2; /root/ceph/run-automount-osd.sh" >  crontab.tmp
+crontab crontab.tmp
+crontab -l
+systemctl enable cront
+systemctl start cront
+#cat  /var/spool/cron/crontabs/root
+
